+++ src/Main/Tests/index.handler.spec.ts
@@ -17,8 +17,11 @@ describe('Start Index Test', () =>
 
     afterAll((async() =>
     {
-        await dbConnection.drop();
-        await dbConnection.close();
+        if (dbConnection)
+        {
+            await dbConnection.drop();
+            await dbConnection.close();
+        }
     }));
 
     describe('#get', () =>
